Backend/requirements.txt
streamlit
python-dotenv
llama-index
llama-index-llms-gemini
mistralai
pypdf
grpcio